Update theme colour and signing
I slightly changed the theme colour to associate with my apps and wanted to update icons accordingly (and also use a more representative icon for the app). Also started string naming the build assemblies. Also fixed an issue where scaled monitors would not show the fortune centered on the screen.